+++ app/Http/Controllers/SupplierController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\IdRequest;
-use App\Http\Requests\SupplierRequest;
+use App\Http\Requests\Supplier\StoreRequest;
+use App\Http\Requests\Supplier\UpdateRequest;
 use App\Models\Supplier;
 use Illuminate\Http\Request;
 
@@ -17,6 +17,8 @@ class SupplierController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', Supplier::class);
+
         $user = $request->user('api');
 
         return $user->suppliers()->with('products')->get();
@@ -25,7 +27,7 @@ class SupplierController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(SupplierRequest $request): Supplier
+    public function store(StoreRequest $request): Supplier
     {
         $user = $request->user('api');
 
@@ -41,6 +43,8 @@ class SupplierController extends Controller
      */
     public function show(Supplier $supplier): Supplier
     {
+        $this->authorize('view', $supplier);
+
         $supplier->load('products');
 
         return $supplier;
@@ -49,7 +53,7 @@ class SupplierController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(SupplierRequest $request, Supplier $supplier): Supplier
+    public function update(UpdateRequest $request, Supplier $supplier): Supplier
     {
         $supplier->fill($request->validated())->save();
 
@@ -58,13 +62,11 @@ class SupplierController extends Controller
 
     /**
      * Remove the specified resource from storage.
-     *
-     *
-     *
-     * @throws \Throwable
      */
     public function destroy(Supplier $supplier): Supplier
     {
+        $this->authorize('delete', $supplier);
+
         $supplier->delete();
 
         return $supplier;
